Added docker container for integration & unit database tests -Moved less unit-y tests to integration -Added extra data types to image -fixed some tests

+++ emonocot-harvest/src/main/java/org/emonocot/job/dwc/image/FieldSetMapper.java
@@ -162,7 +162,7 @@ public class FieldSetMapper extends NonOwnedFieldSetMapper<Image> {
 			XmpTerm xmpTerm = (XmpTerm)term;
 			switch(xmpTerm) {
 			case Rating:
-				object.setRating(conversionService.convert(value, Float.class));
+				object.setRating(conversionService.convert(value, Double.class));
 				break;
 			default:
 				break;
